requirements.txt
python-okx
ratelimit
scikit-learn
redis
vaderSentiment
ta
tensorflow
